backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Config
pydantic-settings==2.6.1
pydantic[email]==2.10.3

# Utils
python-dateutil==2.9.0

# Dev / testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
pytest-cov==6.0.0
aiosqlite==0.20.0
